--- lineapy/_alembic/__init__.py
@@ -1,13 +1,0 @@
-from .alembic_helpers import (
-    maybe_add_column,
-    maybe_create_table,
-    table_exists,
-    table_has_column,
-)
-
-__all__ = [
-    "table_exists",
-    "maybe_create_table",
-    "table_has_column",
-    "maybe_add_column",
-]
